pyproject.toml
[project]
name = "fastapi-mongodb"
version = "0.1.0"
description = ""
authors = [
    {name = "Bruno R. Pianca",email = "brunorpdev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pymongo (>=4.11,<5.0)",
    "python-decouple (>=3.8,<4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
